Reexport strongly typed React and ReactDOM from the window object
This gives access to the global React and ReactDOM objects in a strongly typed context.

The lib file is only included in the build output if it's actually imported and used somewhere, so it shouldn't pollute people's files unless they actually need it.

Unfortunately, i did not find a way to force vscode autoimport to use the local react.ts file instead of the package from node_modules.