Refactor RAM allocation in jackx.js
Reduced the ramBuffer from 512GB to 64GB and updated the conditions for checking available RAM. Now, checks are only made if we have enough RAM to run the script, rather than maintaining a buffer. Also, priority was given to the homeServer by moving it to the start of the server array. These changes improve RAM usage efficiency and processing order.